Fix: don't update release date or mark games as released if they already have dates. Also, don't try to unlink (new) images if the fields aren't set

+++ app/Services/Game/Images.php
@@ -25,6 +25,7 @@ class Images
     {
         $filePath = public_path().self::PATH_IMAGE_SQUARE;
         $fileName = $this->game->image_square;
+        if (!$fileName) return false;
         $fileToDelete = $filePath.$fileName;
         if (file_exists($fileToDelete)) {
             unlink($fileToDelete);
@@ -35,6 +36,7 @@ class Images
     {
         $filePath = public_path().self::PATH_IMAGE_HEADER;
         $fileName = $this->game->image_header;
+        if (!$fileName) return false;
         $fileToDelete = $filePath.$fileName;
         if (file_exists($fileToDelete)) {
             unlink($fileToDelete);
